split up the ui and core logic into separate folders
