+++ src/main/java/org/computational_immunology/ext/ImmuNet/core/StreamedImageServerBuilder.java
@@ -1,4 +1,4 @@
-package org.computational_immunology;
+package org.computational_immunology.ext.ImmuNet.core;
 
 import qupath.lib.images.servers.ImageServer;
 import qupath.lib.images.servers.ImageServerBuilder;
