+++ src/main/resources/META-INF/services/qupath.lib.gui.extensions.QuPathExtension
@@ -1,1 +1,1 @@
-org.computational_immunology.ImmuNetExtension
+org.computational_immunology.ext.ImmuNet.ImmuNetExtension
